+++ java/chat/src/main/java/com/antinvestor/apis/chat/client/ChatClient.java
@@ -2,47 +2,33 @@
 
 package com.antinvestor.apis.chat.client;
 
-import build.buf.gen.chat.v1.ChatServiceClient;
-import build.buf.gen.chat.v1.ChatServiceClientInterface;
-import com.antinvestor.apis.common.base.ConnectClientBase;
+import build.buf.gen.chat.v1.ChatServiceGrpc;
 import com.antinvestor.apis.common.base.GrpcClientBase;
 import com.antinvestor.apis.common.config.DefaultConfig;
 import com.antinvestor.apis.common.context.Context;
-import com.antinvestor.apis.common.interceptor.ClientSideGrpcInterceptor;
-import com.antinvestor.apis.common.utilities.TextUtils;
-import com.antinvestor.apis.common.v1.*;
-import io.grpc.ManagedChannelBuilder;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.*;
 
 /**
  * The ChatClient class represents a client for accessing chat services.
  */
 @ApplicationScoped
-public class ChatClient extends  ConnectClientBase {
+public class ChatClient extends GrpcClientBase<ChatServiceGrpc.ChatServiceBlockingStub> {
 
     @Inject
     public ChatClient(Context context) {
-        var client = setupClient(context);
-
+        setupChannelBuilder(context);
     }
 
     @Override
-    public ConnectionConfig getConnectionConfig(Context context, DefaultConfig defaultConfig) {
+    protected ConnectionConfig getConnectionConfig(Context context, DefaultConfig defaultConfig) {
         var cfg = (ChatConfig) defaultConfig;
-        return  new ConnectionConfig(cfg.notificationsHostUrl(), cfg.notificationsHostPort(), cfg.authInterceptorEnabled() );
+        return new ConnectionConfig(cfg.chatHostUrl(), cfg.chatHostPort(), cfg.authInterceptorEnabled());
     }
 
-    public ChatServiceClientInterface stub(Context context) {
+    public ChatServiceGrpc.ChatServiceBlockingStub stub(Context context) {
 
-        ChatServiceClient
-
-        var stub = NotificationServiceGrpc.newBlockingStub(getChannel());
+        var stub = ChatServiceGrpc.newBlockingStub(getChannel());
         return setupStub(context, stub);
     }
-
 }