+++ java/ledger/src/main/java/com/antinvestor/apis/ledger/client/LedgerClient.java
@@ -3,19 +3,21 @@
 package com.antinvestor.apis.ledger.client;
 
 
+import build.buf.gen.common.v1.SearchRequest;
+import build.buf.gen.ledger.v1.*;
 import com.antinvestor.apis.common.base.GrpcClientBase;
 import com.antinvestor.apis.common.config.DefaultConfig;
 import com.antinvestor.apis.common.context.Context;
 import com.antinvestor.apis.common.exceptions.UnRetriableException;
+import com.antinvestor.apis.common.utilities.IteratorUtil;
 import com.antinvestor.apis.common.utilities.ProtoStructUtil;
-import com.antinvestor.apis.common.v1.SearchRequest;
-import com.antinvestor.apis.ledger.v1.*;
-import com.google.type.Money;
 import io.grpc.ManagedChannel;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 
-import java.util.Iterator;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
+import java.time.format.DateTimeFormatter;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
@@ -49,29 +51,26 @@ public class LedgerClient extends GrpcClientBase<LedgerServiceGrpc.LedgerService
 
         String jsonQuery = String.format("{ \"query\": { \"should\": { \"fields\": [{\"id\": {\"eq\": \"%s\"}}]}}}", reference);
 
-        Iterator<Ledger> ledgerIterator = listLedgers(context, jsonQuery);
-        if (ledgerIterator.hasNext()) {
-            return Optional.of(ledgerIterator.next());
-        } else {
-            return Optional.empty();
-        }
+        Iterable<Ledger> ledgerIterator = listLedgers(context, jsonQuery);
+        return IteratorUtil.firstOf(ledgerIterator);
 
     }
 
-    public Iterator<Ledger> listLedgers(Context context, String jsonQuery) {
+    public Iterable<Ledger> listLedgers(Context context, String jsonQuery) {
         SearchRequest searchRequest = SearchRequest.newBuilder()
                 .setQuery(jsonQuery)
                 .build();
-        return stub(context).searchLedgers(searchRequest);
+        var result = stub(context).searchLedgers(searchRequest);
+        return IteratorUtil.flatMapIterable(result, SearchLedgersResponse::getDataList);
     }
 
     public Ledger createLedger(Context context, String parent, LedgerType type, Map<String, Object> data) {
 
-        Ledger.Builder ledgerBuilder = Ledger.newBuilder()
+        CreateLedgerRequest.Builder ledgerBuilder = CreateLedgerRequest.newBuilder()
                 .setType(type);
 
         if (!Objects.isNull(parent)) {
-            ledgerBuilder.setParent(parent);
+            ledgerBuilder.setParentId(parent);
         }
 
         if (!Objects.isNull(data)) {
@@ -79,15 +78,20 @@ public class LedgerClient extends GrpcClientBase<LedgerServiceGrpc.LedgerService
             ledgerBuilder.setData(ProtoStructUtil.fromMap(data));
         }
 
-        return stub(context).createLedger(ledgerBuilder.build());
+        var result = stub(context).createLedger(ledgerBuilder.build());
+        return result.getData();
     }
 
     public Optional<Ledger> updateLedger(Context context, String reference, Map<String, Object> data) {
         Optional<Ledger> optionalLedger = getLedger(context, reference);
         if (optionalLedger.isPresent()) {
-            var ledger = optionalLedger.get();
-            ledger = ledger.toBuilder().setData(ProtoStructUtil.fromMap(data)).build();
-            return Optional.of(stub(context).updateLedger(ledger));
+            var updateLedgerRequest = UpdateLedgerRequest.newBuilder()
+                    .setId(reference)
+                    .setData(ProtoStructUtil.fromMap(data))
+                    .build();
+            var result = stub(context).updateLedger(updateLedgerRequest);
+
+            return Optional.of(result.getData());
         } else {
             return Optional.empty();
         }
@@ -97,43 +101,41 @@ public class LedgerClient extends GrpcClientBase<LedgerServiceGrpc.LedgerService
     public Optional<Account> getAccount(Context context, String reference) {
 
         String jsonQuery = String.format("{ \"query\": { \"should\": { \"fields\": [{\"id\": {\"eq\": \"%s\"}}]}}}", reference);
-        Iterator<Account> accountIterator = listAccounts(context, jsonQuery);
-        if (accountIterator.hasNext()) {
-            return Optional.of(accountIterator.next());
-        } else {
-            return Optional.empty();
-        }
-
+        Iterable<Account> accountIterator = listAccounts(context, jsonQuery);
+        return IteratorUtil.firstOf(accountIterator);
     }
 
-    public Iterator<Account> listAccounts(Context context, String jsonQuery) {
+    public Iterable<Account> listAccounts(Context context, String jsonQuery) {
         SearchRequest searchRequest = SearchRequest.newBuilder()
                 .setQuery(jsonQuery)
                 .build();
-        return stub(context).searchAccounts(searchRequest);
+        var result = stub(context).searchAccounts(searchRequest);
+
+        return IteratorUtil.flatMapIterable(result, SearchAccountsResponse::getDataList);
     }
 
     public Account createAccount(Context context, String ledgerReference, String currencyCode, Map<String, Object> data) {
 
-        Money money = Money.newBuilder().setUnits(0).setNanos(0).setCurrencyCode(currencyCode).build();
-        Account.Builder accountBuilder = Account.newBuilder()
-                .setBalance(money)
-                .setLedger(ledgerReference);
+        CreateAccountRequest.Builder accountBuilder = CreateAccountRequest.newBuilder()
+                .setCurrency(currencyCode)
+                .setLedgerId(ledgerReference);
 
         if (!Objects.isNull(data)) {
             accountBuilder.setData(ProtoStructUtil.fromMap(data));
         }
 
-        return stub(context).createAccount(accountBuilder.build());
+        var result = stub(context).createAccount(accountBuilder.build());
+        return result.getData();
     }
 
     public Optional<Account> updateAccount(Context context, String reference, Map<String, Object> data) {
 
-        Account.Builder accountBuilder = Account.newBuilder()
-                .setReference(reference)
+        UpdateAccountRequest.Builder accountBuilder = UpdateAccountRequest.newBuilder()
+                .setId(reference)
                 .setData(ProtoStructUtil.fromMap(data));
 
-        return Optional.of(stub(context).updateAccount(accountBuilder.build()));
+        var response = stub(context).updateAccount(accountBuilder.build());
+        return Optional.of(response.getData());
     }
 
 
@@ -153,29 +155,32 @@ public class LedgerClient extends GrpcClientBase<LedgerServiceGrpc.LedgerService
         SearchRequest searchRequest = SearchRequest.newBuilder()
                 .setQuery(jsonQuery)
                 .build();
-        Iterator<Transaction> results = stub(context).searchTransactions(searchRequest);
-        return () -> results;
+        var result = stub(context).searchTransactions(searchRequest);
+        return IteratorUtil.flatMapIterable(result, SearchTransactionsResponse::getDataList);
     }
 
     public Transaction createTransaction(Context context, TransactionBuilder transactionBuilder) throws UnRetriableException {
 
-        Transaction transaction = transactionBuilder.build();
+        CreateTransactionRequest transaction = transactionBuilder.build();
 
-        return stub(context).createTransaction(transaction);
+        var result = stub(context).createTransaction(transaction);
+        return result.getData();
     }
 
     public Optional<Transaction> clearTransaction(Context context, String reference) {
         Optional<Transaction> optionalTransaction = getTransaction(context, reference);
 
-        if (optionalTransaction.isPresent()) {
-            var transaction = optionalTransaction.get();
-            Transaction.Builder transactionBuilder = Transaction.newBuilder(transaction);
-            transactionBuilder.setCleared(true);
-            var nueTransaction = transactionBuilder.build();
-            return Optional.of(stub(context).updateTransaction(nueTransaction));
-        } else {
+        if (optionalTransaction.isEmpty()) {
             return Optional.empty();
         }
+            var transaction = optionalTransaction.get();
+            UpdateTransactionRequest.Builder transactionBuilder = UpdateTransactionRequest.newBuilder();
+            transactionBuilder.setId(transaction.getId());
+            transactionBuilder.setClearedAt(LocalDateTime.now(ZoneOffset.UTC).format(DateTimeFormatter.ISO_DATE_TIME));
+            var nueTransaction = transactionBuilder.build();
+            var result = stub(context).updateTransaction(nueTransaction);
+            return Optional.of(result.getData());
+
     }
 
     public Optional<Transaction> updateTransaction(Context context, String reference, Map<String, Object> data) {
@@ -183,10 +188,12 @@ public class LedgerClient extends GrpcClientBase<LedgerServiceGrpc.LedgerService
 
         if (optionalTransaction.isPresent()) {
 
-            Transaction.Builder transactionBuilder = Transaction.newBuilder(optionalTransaction.get());
+            UpdateTransactionRequest.Builder transactionBuilder = UpdateTransactionRequest.newBuilder();
+            transactionBuilder.setId(optionalTransaction.get().getId());
             transactionBuilder.setData(ProtoStructUtil.fromMap(data));
-            Transaction transaction = transactionBuilder.build();
-            return Optional.of(stub(context).updateTransaction(transaction));
+            UpdateTransactionRequest transaction = transactionBuilder.build();
+            var result = stub(context).updateTransaction(transaction);
+            return Optional.of(result.getData());
         } else {
             return Optional.empty();
         }
@@ -195,9 +202,11 @@ public class LedgerClient extends GrpcClientBase<LedgerServiceGrpc.LedgerService
 
     public Optional<Transaction> reverseTransaction(Context context, String reference) {
 
-        Transaction transaction = Transaction.newBuilder()
-                .setReference(reference).build();
-        return Optional.ofNullable(stub(context).reverseTransaction(transaction));
+        ReverseTransactionRequest transaction = ReverseTransactionRequest.newBuilder()
+                .setId(reference).build();
+
+        var response = stub(context).reverseTransaction(transaction);
+        return Optional.of(response.getData());
     }
 
 
@@ -205,8 +214,8 @@ public class LedgerClient extends GrpcClientBase<LedgerServiceGrpc.LedgerService
         SearchRequest searchRequest = SearchRequest.newBuilder()
                 .setQuery(jsonQuery)
                 .build();
-        Iterator<TransactionEntry> results = stub(context).searchTransactionEntries(searchRequest);
-        return () -> results;
+        var results = stub(context).searchTransactionEntries(searchRequest);
+        return IteratorUtil.flatMapIterable(results, SearchTransactionEntriesResponse::getDataList);
     }
 
 }